backend/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "portfolio-backend"
version = "1.0.0"
description = "Backend for the Portfolio AI Assistant."
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "python-multipart==0.0.6",
    "google-generativeai==0.3.1",
    "openai==1.10.0",
    "python-dotenv==1.0.0",
    "pydantic>=2.0.0",
    "email-validator==2.1.0",
    "requests==2.31.0",
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "ruff",
]

[tool.setuptools]
packages = ["agents", "routes"]

[tool.ruff]
line-length = 88
